+++ all-is-cubes-port/src/mv/import.rs
@@ -188,7 +188,7 @@ pub(crate) async fn dot_vox_data_to_universe(
         .insert(
             "character".into(),
             Character::spawn_default(universe.read_ticket(), viewed_space_handle)
-                .map_err(|e| DotVoxConversionError::Unexpected(InGenError::other(e)))?,
+                .map_err(|e| DotVoxConversionError::Unexpected(InGenError::Handle(e)))?,
         )
         .map_err(|e| DotVoxConversionError::Unexpected(InGenError::from(e)))?;